requirements.txt
colorama==0.4.5
python-dotenv==1.0.0
selenium==4.13.0
